+++ oalg-base/src/OAlg/Control/Validate.hs
@@ -167,7 +167,7 @@ import OAlg.Category.Definition
 import OAlg.Control.Verbose
 import OAlg.Control.HNFData
 
-import OAlg.Data.Ord.Definition
+import OAlg.Data.Ord
 import OAlg.Data.Statement
 
 import OAlg.Data.Boolean
