+++ oalg-base/src/OAlg/Proposition.hs
@@ -15,6 +15,7 @@ module OAlg.Proposition
 
 import OAlg.Prelude
 import OAlg.Structure.Proposition
+import OAlg.Structure.Lattice.Proposition
 import OAlg.Hom.Proposition
 import OAlg.Limes.Proposition
 
@@ -33,7 +34,7 @@ prpOAlgBase = Prp "OAlgBase"
                 And [ prpBool
                     , prpValidTautologies
                     , prpStatement
-                    , prpOrd
+                    , prpLatticeBool
                     ]
             , Label "Structure" :<=>:
                 And [ prpStructureN
