+++ oalg-base/src/OAlg/Structure/Lattice.hs
@@ -1,0 +1,17 @@
+
+-- |
+-- Module      : OAlg.Structure.Lattice
+-- Description : lattices.
+-- Copyright   : (c) Erich Gut
+-- License     : BSD3
+-- Maintainer  : zerich.gut@gmail.com
+--
+-- 'Lattice's.
+module OAlg.Structure.Lattice
+  ( module Ltc
+  , module Prp
+  )
+  where
+
+import OAlg.Structure.Lattice.Definition as Ltc
+import OAlg.Structure.Lattice.Proposition as Prp
